Add implementation for MonitoringServer
This was a bit rushed and there is some duplication especially in the
DoubleKeyedMap. There also *must* be a better way to do that, it's all a
bit convoluted. Tests are present, but a bit sparse as well.

That said this will be integrated into the MonitorDispachProtocol
and MonitorKeepalive server.

All combined should form a decent interface to forward data from the
launched processes to external monitoring devices.